Deduplicate oAuth2client config code
